+++ gdsfactory/components/nxn.py
@@ -141,5 +141,5 @@ def nxn(
 
 if __name__ == "__main__":
     c = nxn(north=1.3, south=3)
-    # c = gf.extend.extend_ports(component=c)
+    # c = gf.components.extension.extend_ports(component=c)
     c.show()